requirements.txt
accelerate>=1.0.0
agno>=2.4.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0
tzlocal>=5.0.0
annotated-types>=0.6.0
antlr4-python3-runtime>=4.9.0
anyio>=4.0.0
asgiref>=3.7.0
attrs>=23.0.0
beautifulsoup4>=4.12.0
bleach>=6.0.0
certifi>=2024.0.0
cffi>=1.15.0
charset-normalizer>=3.0.0
click>=8.1.0
colorlog>=6.7.0
cryptography>=42.0.0
deprecation>=2.1.0
dill>=0.3.0
distro>=1.8.0
Django>=4.2.0
django-picklefield>=3.1.0
django-q2>=1.6.0
docstring_parser>=0.15.0
et_xmlfile>=1.1.0
Faker>=20.0.0
filelock>=3.12.0
filetype>=1.2.0
fsspec>=2024.0.0
gitdb>=4.0.0
GitPython>=3.1.0
google-api-core>=2.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.0.0
googleapis-common-protos>=1.60.0
greenlet>=3.0.0
h11>=0.14.0
h2>=4.0.0
hf-xet>=1.0.0
hpack>=4.0.0
httpcore>=1.0.0
httplib2>=0.22.0
httpx>=0.27.0
huggingface_hub>=0.20.0
hyperframe>=6.0.0
idna>=3.4
Jinja2>=3.1.0
jiter>=0.5.0
jsonlines>=4.0.0
jsonpatch>=1.32
jsonpointer>=2.4
jsonref>=1.1.0
jsonschema>=4.20.0
jsonschema-specifications>=2023.0.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.1.0
langchain-openai>=0.1.0
langchain-text-splitters>=0.2.0
langsmith>=0.1.0
faiss-cpu>=1.7.0
latex2mathml>=3.75.0
lxml>=5.0.0
Markdown>=3.5.0
markdown-it-py>=3.0.0
marko>=2.0.0
MarkupSafe>=2.1.0
martor>=1.6.0
mdurl>=0.1.2
mpire>=2.7.0
mpmath>=1.3.0
multiprocess>=0.70.0
networkx>=3.0.0
numpy>=1.24.0,<2.0.0
oauthlib>=3.2.0
omegaconf>=2.3.0
openai>=1.0.0
opencv-python>=4.8.0
openpyxl>=3.1.0
orjson>=3.9.0
packaging>=23.0
pandas>=2.0.0
pillow>=10.0.0
pluggy>=1.3.0
polyfactory>=2.0.0
proto-plus>=1.22.0
protobuf>=4.0.0,<5.0.0
psutil>=5.9.0
pyarrow>=15.0.0
pyasn1>=0.5.0
pyasn1_modules>=0.3.0
pyclipper>=1.3.0
pycparser>=2.21
pydantic>=2.0.0
pydantic-settings>=2.0.0
pydantic_core>=2.0.0
Pygments>=2.15.0
pylatexenc>=2.10
pyparsing>=3.0.0
pypdfium2>=4.0.0
pdf2image>=1.16.0
python-dateutil>=2.8.0
python-docx>=1.1.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
python-pptx>=0.6.0
pytz>=2024.0
PyYAML>=6.0
rapidocr>=3.0.0
referencing>=0.30.0
regex>=2024.0.0
requests>=2.31.0
requests-oauthlib>=1.3.0
requests-toolbelt>=1.0.0
rich>=13.0.0
rpds-py>=0.10.0
rsa>=4.9.0
rtree>=1.0.0
safetensors>=0.4.0
scipy>=1.10.0
setuptools>=68.0.0
shapely>=2.0.0
shellingham>=1.5.0
six>=1.16.0
smmap>=5.0.0
sniffio>=1.3.0
soupsieve>=2.5
SQLAlchemy>=2.0.0
sqlparse>=0.4.0
sympy>=1.12
tabulate>=0.9.0
tenacity>=8.0.0
tiktoken>=0.5.0
tokenizers>=0.13.0,<0.20.0
torch>=2.0.0,<3.0.0
torchvision>=0.15.0
tqdm>=4.65.0
transformers>=4.30.0
tree-sitter>=0.20.0
tree-sitter-c>=0.20.0
tree-sitter-javascript>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-typescript>=0.20.0
typer>=0.9.0
typing_extensions>=4.7.0
tzdata>=2024.0
tzlocal>=5.0.0
uritemplate>=4.1.0
urllib3>=2.0.0
uuid_utils>=0.9.0
webencodings>=0.5.1
xlsxwriter>=3.1.0
xxhash>=3.0.0
zipp>=3.15.0
zstandard>=0.21.0
